removing unnecessary state from wallet/index into AddWalletModal
